Further seperate to more object files
