+++ cpp/Osmosis/Chain/Local/ObjectStore.h
@@ -1,8 +1,12 @@
 #ifndef __OSMOSIS_CHAIN_LOCAL_OBJECT_STORE_H__
 #define __OSMOSIS_CHAIN_LOCAL_OBJECT_STORE_H__
 
+#include <boost/filesystem.hpp>
 #include "Osmosis/Chain/ObjectStoreInterface.h"
 #include "Osmosis/Chain/Local/Connection.h"
+#include "Osmosis/ObjectStore/Store.h"
+#include "Osmosis/ObjectStore/Drafts.h"
+#include "Osmosis/ObjectStore/Labels.h"
 
 namespace Osmosis {
 namespace Chain {
@@ -12,16 +16,9 @@ namespace Local
 class ObjectStore : public ObjectStoreInterface
 {
 public:
-	ObjectStore( const boost::filesystem::path & path ) :
-		_store( path ),
-		_drafts( path ),
-		_labels( path, _store )
-	{}
+	ObjectStore( const boost::filesystem::path & path );
 
-	std::unique_ptr< ObjectStoreConnectionInterface > connect() override
-	{
-		return std::unique_ptr< ObjectStoreConnectionInterface >( new Connection( _store, _drafts, _labels ) );
-	}
+	std::unique_ptr< ObjectStoreConnectionInterface > connect() override;
 
 private:
 	Osmosis::ObjectStore::Store   _store;
