+++ cpp/Osmosis/Client/DigestDrafts.h
@@ -1,6 +1,8 @@
 #ifndef __OSMOSIS_CLIENT_DIGEST_DRAFTS_H__
 #define __OSMOSIS_CLIENT_DIGEST_DRAFTS_H__
 
+#include <thread>
+#include "Osmosis/CalculateHash.h"
 #include "Osmosis/Client/Typedefs.h"
 #include "Common/NumberOfCPUs.h"
