+++ cpp/Osmosis/Stream/CopyFile.h
@@ -11,34 +11,15 @@ namespace Stream
 class CopyFile
 {
 public:
-	CopyFile( const char * fromPath, const char * toPath ) :
-		_read( fromPath ),
-		_write( toPath )
-	{}
+	CopyFile( const char * fromPath, const char * toPath );
 
-	void copy()
-	{
-		bool lastReadWasHole = false;
-		while ( not _read.done() ) {
-			lastReadWasHole = _read.guessHole();
-			if ( not lastReadWasHole )
-				_write.write( _read.offset(), _read.buffer(), _read.length() );
-			_read.next();
-		}
-		if ( lastReadWasHole )
-			writeZeroCharacterAtLastOffset();
-	}
+	void copy();
 
 private:
 	ReadFile   _read;
 	WriteFile  _write;
 
-	void writeZeroCharacterAtLastOffset()
-	{
-		unsigned char zero = 0;
-		ASSERT( _read.totalLength() >= 1 );
-		_write.write( _read.totalLength() - sizeof( zero ), & zero, sizeof( zero ) );
-	}
+	void writeZeroCharacterAtLastOffset();
 
 	CopyFile( const CopyFile & rhs ) = delete;
 	CopyFile & operator= ( const CopyFile & rhs ) = delete;
