+++ cpp/Osmosis/main.cpp
@@ -7,6 +7,7 @@
 #include "Osmosis/ObjectStore/LabelLogIterator.h"
 #include "Osmosis/ObjectStore/LeastRecentlyUsed.h"
 #include "Osmosis/ObjectStore/Purge.h"
+#include "Osmosis/FilesystemUtils.h"
 
 std::mutex globalTraceLock;
